Move to builtin_controller.rs
